% Generated by roxygen2 (4.0.0): do not edit by hand
\name{abs_area}
\alias{abs_area}
\title{Point area palette (continuous), with area proportional to value.}
\usage{
abs_area(max)
}
\arguments{
\item{max}{A number representing the maxmimum size.}
}
\description{
Point area palette (continuous), with area proportional to value.
}

