% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pal-area.r
\name{area_pal}
\alias{area_pal}
\title{Point area palette (continuous).}
\usage{
area_pal(range = c(1, 6))
}
\arguments{
\item{range}{Numeric vector of length two, giving range of possible sizes.
Should be greater than 0.}
}
\description{
Point area palette (continuous).
}

