% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans-numeric.r
\name{asinh_trans}
\alias{asinh_trans}
\title{Inverse Hyperbolic Sine transformation}
\usage{
asinh_trans()
}
\description{
Inverse Hyperbolic Sine transformation
}
\examples{
plot(asinh_trans(), xlim = c(-1e2, 1e2))
}
