\name{brewer_pal}
\alias{brewer_pal}
\title{Color Brewer palette (discrete).}
\usage{brewer_pal(type="seq", palette=1)}

\description{
  Color Brewer palette (discrete).
}
\references{\url{http://colorbrewer2.org}}
\arguments{
  \item{type}{One of seq (sequential), div (diverging) or qual (qualitative)}
  \item{palette}{If a string, will use that named palette.  If a number, will
index into the list of palettes of appropriate \code{type}}
}
