% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatter.r
\name{byte_format}
\alias{byte_format}
\alias{bytes}
\title{Bytes formatter: convert to byte measurement and display symbol.}
\usage{
byte_format(symbol = "auto", units = "binary", ...)

bytes(x, symbol = "auto", units = c("binary", "si"), ...)
}
\arguments{
\item{units}{which unit base to use, "binary" (1024 base) or
"si" (1000 base) for ISI units.}

\item{...}{other arguments passed to \code{\link[=number]{number()}}}

\item{x}{a numeric vector to format}
}
\value{
a function with three parameters, \code{x``, a numeric vector that returns a character vector,}symbol\code{the byte symbol (e.g. "Kb") desired and the measurement}units\code{(traditional}binary\code{or}si` for ISI metric units).
}
\description{
Bytes formatter: convert to byte measurement and display symbol.
}
\examples{
byte_format()(sample(3000000000, 10))
bytes(sample(3000000000, 10))
bytes(sample(3000000000, 10), accuracy = .1)
}
\references{
Units of Information (Wikipedia) :
\url{http://en.wikipedia.org/wiki/Units_of_information}
}
