% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.r
\name{calculate_minor_breaks}
\alias{calculate_minor_breaks}
\title{Minor breaks.
Places minor breaks between major breaks and within limits.}
\usage{
calculate_minor_breaks(reverse = FALSE)
}
\arguments{
\item{reverse}{if TRUE, calculates the minor breaks for a reversed scale}
}
\description{
Minor breaks.
Places minor breaks between major breaks and within limits.
}
\examples{
m <- extended_breaks()(1:10)
calculate_minor_breaks()(m, 1:10, n = 2)

n <- extended_breaks()(0:-9)
calculate_minor_breaks(reverse = TRUE)(n, 0:-9, n = 2)
}
