\name{censor}
\alias{censor}
\title{Censor any values outside of range.}
\usage{censor(x, range=c(0, 1))}

\description{
  Censor any values outside of range.
}
\arguments{
  \item{x}{numeric vector of values to manipulate.}
  \item{range}{numeric vector of length two giving desired output range.}
}
