\name{clip}
\alias{clip}
\title{Clip values to range.}
\usage{clip(x, range=c(0, 1))}

\description{
  Clip values to range.
}
\author{Homer Strong <homer.strong@gmail.com>}
\arguments{
  \item{x}{numeric vector of values to manipulate.}
  \item{range}{numeric vector of length two giving desired output range.}
}
