\name{clip_infinite}
\alias{clip_infinite}
\title{Clip infinite values to range.}
\usage{clip_infinite(x, range=c(0, 1))}

\description{
  Clip infinite values to range.
}
\arguments{
  \item{x}{numeric vector of values to manipulate.}
  \item{range}{numeric vector of length two giving desired output range.}
}
