% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour-manip.R
\name{col_contrasting}
\alias{col_contrasting}
\title{Choose a contrasting colour}
\usage{
col_contrasting(colour, light = "white", dark = "black")
}
\arguments{
\item{colour}{character vector of colours to be modified}

\item{light}{colour to return when \code{colour} is dark}

\item{dark}{colour to return when \code{colour} is light}
}
\value{
A character vector of colours with the same length as \code{colour}
}
\description{
Chooses either \code{light} or \code{dark} based on the lightness of \code{colour}.
This is useful for choosing a text colour that is legible on a solid background
cooured with \code{colour}.
}
\examples{
col_contrasting(c("navy", "white", "black", "yellow"))

}
\seealso{
Other colour manipulation: 
\code{\link{alpha}()},
\code{\link{col2hcl}()},
\code{\link{col_mix}()},
\code{\link{colour_manip}},
\code{\link{muted}()}
}
\concept{colour manipulation}
