% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatter.r
\name{comma_format}
\alias{comma_format}
\alias{comma}
\title{Comma formatter: format number with commas separating thousands.}
\usage{
comma_format(big.mark = ",", ...)

comma(x, big.mark = ",", ...)
}
\arguments{
\item{big.mark}{character used between every 3 digits to separate thousands}

\item{...}{other arguments passed on to \code{\link{number_format}}}

\item{x}{a numeric vector to format}
}
\value{
a function with single parameter x, a numeric vector, that
returns a character vector
}
\description{
Comma formatter: format number with commas separating thousands.
}
\examples{
comma_format()(c(1, 1e3, 2000, 1e6))
comma_format(accuracy = .1)(c(1, 1e3, 2000, 1e6))
comma(c(1, 1e3, 2000, 1e6))

# If you're European you can switch . and , with 'big.mark'
# and 'decimal.mark' arguments (see number_format)
point <- comma_format(big.mark = ".", decimal.mark = ",", accuracy = .01)
point(c(1, 1e3, 2000, 1e6))
point(c(1, 1.021, 1000.01))
}
