% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-compose.R
\name{compose_label}
\alias{compose_label}
\title{Compose two or more label formatters together}
\usage{
compose_label(..., call = caller_env())
}
\arguments{
\item{...}{One or more labelling functions. These will be applied to breaks
consecutively.
\link[rlang:as_function]{Lambda syntax} is allowed.}

\item{call}{A call to display in error messages.}
}
\value{
A labelling function that applies the provided
functions to breaks to return labels.
}
\description{
This labeller provides a general mechanism for composing two or more
labellers together.
}
\examples{
demo_continuous(
  c(-100, 100),
  labels = compose_label(abs, number, ~paste0(.x, " foobar"), toupper)
)

# Same result
demo_continuous(
  c(-100, 100),
  labels = compose_label(abs, label_number(suffix = " FOOBAR"))
)
}
