% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatter.r
\name{coordinate_format}
\alias{coordinate_format}
\alias{latitude_format}
\alias{latitude}
\alias{longitude_format}
\alias{longitude}
\title{Format numbers as geographic coordinates}
\usage{
coordinate_format(axis = c("latitude", "longitude"), accuracy = 1,
  scale = 1, prefix = "", sep = "", big.mark = " ",
  decimal.mark = ".", trim = TRUE, ...)

latitude_format(accuracy = 1, scale = 1, prefix = "", sep = "",
  big.mark = " ", decimal.mark = ".", trim = TRUE, ...)

latitude(x, accuracy = 1, scale = 1, prefix = "", sep = "",
  big.mark = " ", decimal.mark = ".", trim = TRUE, ...)

longitude_format(accuracy = 1, scale = 1, prefix = "", sep = "",
  big.mark = " ", decimal.mark = ".", trim = TRUE, ...)

longitude(x, accuracy = 1, scale = 1, prefix = "", sep = "",
  big.mark = " ", decimal.mark = ".", trim = TRUE, ...)
}
\arguments{
\item{axis}{String specifying the coordinate axis. Either "latitude" or
"longitude".}

\item{accuracy}{Number to round to, \code{NULL} for automatic guess.}

\item{scale}{A scaling factor: \code{x} will be multiply by \code{scale} before
formating (useful if the underlying data is on another scale,
e.g. for computing percentages or thousands).}

\item{prefix}{Symbols to display before and after value.}

\item{sep}{The separator between the number and the unit label.}

\item{big.mark}{Character used between every 3 digits to separate thousands.}

\item{decimal.mark}{The character to be used to indicate the numeric
decimal point.}

\item{trim}{Logical, if \code{FALSE}, values are right-justified to a common
width (see \code{\link[base:format]{base::format()}}).}

\item{...}{Other arguments passed on to \code{\link[base:format]{base::format()}}.}

\item{x}{A numeric vector to format.}
}
\description{
Format numbers as geographic coordinates
}
\examples{
latitude(seq(-20, 20, by = 10))
longitude(seq(-20, 20, by = 10))
}
\seealso{
\code{\link[=degree_format]{degree_format()}} that powers these formatters
}
