% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{demo_continuous}
\alias{demo_continuous}
\alias{demo_log10}
\alias{demo_discrete}
\title{Demonstrate scales functions with ggplot2 code}
\usage{
demo_continuous(x, ...)

demo_log10(x, ...)

demo_discrete(x, ...)
}
\arguments{
\item{x}{A vector of data}
}
\description{
These functions generate ggplot2 code needed to use scales functions for
real code.
}
\keyword{internal}
