\name{discard}
\alias{discard}
\title{Discard any values outside of range.}
\usage{discard(x, range=c(0, 1))}

\description{
  Discard any values outside of range.
}
\arguments{
  \item{x}{numeric vector of values to manipulate.}
  \item{range}{numeric vector of length two giving desired output range.}
}
