% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans-numeric.r
\name{exp_trans}
\alias{exp_trans}
\title{Exponential transformation (inverse of log transformation)}
\usage{
exp_trans(base = exp(1))
}
\arguments{
\item{base}{Base of logarithm}
}
\description{
Exponential transformation (inverse of log transformation)
}
\examples{
plot(exp_trans(0.5), xlim = c(-2, 2))
plot(exp_trans(1), xlim = c(-2, 2))
plot(exp_trans(2), xlim = c(-2, 2))
plot(exp_trans(), xlim = c(-2, 2))
}
