% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bounds.r
\name{expand_range}
\alias{expand_range}
\title{Expand a range with a multiplicative or additive constant.}
\usage{
expand_range(range, mul = 0, add = 0, zero_width = 1)
}
\arguments{
\item{range}{range of data, numeric vector of length 2}

\item{mul}{multiplicative constract}

\item{add}{additive constant}

\item{zero_width}{distance to use if range has zero width}
}
\description{
Expand a range with a multiplicative or additive constant.
}

