\name{fullseq}
\alias{fullseq}
\title{Generate sequence of fixed size intervals covering range.}
\usage{fullseq(range, size, ...)}

\description{
  Generate sequence of fixed size intervals covering range.
}
\keyword{internal}
\seealso{\code{\link{plyr}{round_any}}}
\arguments{
  \item{range}{range}
  \item{size}{interval size}
  \item{...}{other arguments passed on to methods}
}
