% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-registry.R
\name{get_palette}
\alias{get_palette}
\alias{set_palette}
\alias{palette_names}
\alias{reset_palettes}
\title{Known palettes}
\usage{
get_palette(name, ...)

set_palette(name, palette, warn_conflict = TRUE)

palette_names()

reset_palettes()
}
\arguments{
\item{name}{A string giving the palette name.}

\item{...}{Additional arguments to pass to palette when it is a function
but not a palette class function.}

\item{palette}{A \link[=new_continuous_palette]{palette}, \code{function} or character
vector.}

\item{warn_conflict}{A boolean which if \code{TRUE} (default), warns when
replacing a known palette.}
}
\value{
The \code{get_palette()} function returns a palette. The \code{set_palette()}
function is called for side effects and returns nothing.
}
\description{
The scales packages keeps track of a set of palettes it considers 'known'.
The benefit of a known palette is that it can be called by name in functions
as \code{as_continuous_pal()} or \code{as_discrete_pal()}.
}
\examples{
# Get one of the known palettes
get_palette("Okabe-Ito")

# Set a new custom palette
cols <- c("palegreen", "deepskyblue", "magenta")
set_palette("aurora", palette = cols)

# Palette is now known
"aurora" \%in\% palette_names()
as_continuous_pal("aurora")

# Resetting palettes
reset_palettes()
}
