\name{gradient_n_pal}
\alias{gradient_n_pal}
\title{Arbitrary colour gradient palette (continous).}
\usage{gradient_n_pal(colours, values, space="Lab")}

\description{
  Arbitrary colour gradient palette (continous).
}
\arguments{
  \item{colours}{vector of colours}
  \item{values}{if colours should not be evenly positioned along the gradient}
  \item{space}{colour space in which to calculate gradient.  "Lab" usually
best unless gradient goes through white.}
}
