% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans-date.r
\name{hms_trans}
\alias{hms_trans}
\title{Transformation for times (class hms).}
\usage{
hms_trans()
}
\description{
Transformation for times (class hms).
}
\examples{
if (require("hms")) {
hms <- round(runif(10) * 86400)
t <- hms_trans()
t$transform(hms)
t$inverse(t$transform(hms))
t$breaks(hms)
}
}
