% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-itemized.R
\name{label_itemized}
\alias{label_itemized}
\title{Label categories like an itemized list}
\usage{
label_itemized(fmt = "a")
}
\arguments{
\item{fmt}{A format string containing exactly one occurrence of either \code{"i"},
\code{"I"}, \code{"a"}, \code{"A"} or \code{"1"}.}
}
\value{
All \code{label_()} functions return a "labelling" function, i.e. a function that
takes a vector \code{x} and returns a character vector of \code{length(x)} giving a
label for each input value.

Labelling functions are designed to be used with the \code{labels} argument of
ggplot2 scales. The examples demonstrate their use with x scales, but
they work similarly for all scales, including those that generate legends
rather than axes.
}
\description{
\code{label_itemized} displays values using itemized list numbering
like \verb{a, b, c} or \verb{(1), (2), (3)}.
}
\details{
Values are converted to integers.
\itemize{
\item If \code{fmt} contains \code{"a"}, values are labelled \verb{a, b, c, ...}.
\item If \code{fmt} contains \code{"A"},  values are labelled \verb{A, B, C, ...}.
\item If \code{fmt} contains \code{"i"},  values are labelled by Roman numerals
\verb{i, ii, iii, iv, ...}.
\item If \code{fmt} contains \code{"I"},  values are labelled \verb{I, II, III, IV, ...}.
\item If \code{fmt} contains \code{"1"}, values are labelled \verb{1, 2, 3,...}.
}

Other characters in \code{fmt} are passed through as-is, so e.g. \code{fmt = "a)"}
becomes \verb{a), b), c), ...}.
}
\examples{
demo_continuous(1:5)
demo_continuous(1:5, labels = label_itemized())
demo_continuous(1:5, labels = label_itemized("(a)"))
demo_continuous(1:5, labels = label_itemized("i."))
}
