% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits.R
\name{limits_center}
\alias{limits_center}
\title{Centered limits}
\usage{
limits_center(center = 0)
}
\arguments{
\item{center}{A scalar numeric value to serve as midpoint.}
}
\value{
The \code{limits_()} functions return a function for setting limits.
These functions take as their argument a vector values that represent the
natural data limits.
}
\description{
Centered limits
}
\examples{
# Symmetry around 0
demo_continuous(c(-5, 10), limits = limits_center(0))
}
