% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits.R
\name{limits_include}
\alias{limits_include}
\title{Limits include value}
\usage{
limits_include(value = 0)
}
\arguments{
\item{value}{A scalar numeric value to include.}
}
\value{
The \code{limits_()} functions return a function for setting limits.
These functions take as their argument a vector values that represent the
natural data limits.
}
\description{
Limits include value
}
\examples{
# Ensure 0 is included in the scale range
demo_continuous(c(5, 10), limits = limits_include(0))
}
