% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans-numeric.r
\name{log1p_trans}
\alias{log1p_trans}
\title{Log plus one transformation}
\usage{
log1p_trans()
}
\description{
Log plus one transformation
}
\examples{
plot(log1p_trans(), xlim = c(-1, 1))
# Note the x-axis
plot(log_trans(), xlim = c(0, 2))

trans_range(log_trans(), 1:10)
trans_range(log1p_trans(), 0:9)
}
