% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans-numeric.r
\name{log_trans}
\alias{log_trans}
\alias{log10_trans}
\alias{log2_trans}
\title{Log transformation.}
\usage{
log_trans(base = exp(1))

log10_trans()

log2_trans()
}
\arguments{
\item{base}{base of logarithm}
}
\description{
Log transformation.
}
\examples{
plot(log2_trans(), xlim = c(0, 5))
plot(log_trans(), xlim = c(0, 5))
plot(log10_trans(), xlim = c(0, 5))
}
