% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-manual.r
\name{manual_pal}
\alias{manual_pal}
\title{Manual palette (discrete)}
\usage{
manual_pal(values)
}
\arguments{
\item{values}{vector of values to be used as a palette.}
}
\description{
Manual palette (discrete)
}
