\name{map_continuous}
\alias{map_continuous}
\title{Map values for a continuous palette.}
\usage{map_continuous(palette, x, limits, na.value=NA, oob=censor)}

\description{
  Map values for a continuous palette.
}
\arguments{
  \item{oob}{out of bounds behaviour. Defaults to \code{\link{censor}}
which turns oob values into missing values.}
}
