% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.r
\name{minor_breaks}
\alias{minor_breaks}
\title{Minor breaks.
Places minor breaks between major breaks.}
\usage{
minor_breaks(reverse = FALSE)
}
\arguments{
\item{reverse}{if TRUE, calculates the minor breaks for a reversed scale}
}
\description{
Minor breaks.
Places minor breaks between major breaks.
}
\examples{
m <- extended_breaks()(c(1, 10))
minor_breaks()(m, c(1, 10), n = 2)

n <- extended_breaks()(c(0, -9))
minor_breaks(reverse = TRUE)(n, c(0, -9), n = 2)
}
