% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minor_breaks.R
\name{minor_breaks_width}
\alias{minor_breaks_width}
\alias{minor_breaks_n}
\title{Minor breaks}
\usage{
minor_breaks_width(width, offset)

minor_breaks_n(n)
}
\arguments{
\item{width}{Distance between each break. Either a number, or for
date/times, a single string of the form \code{"{n} {unit}"}, e.g. "1 month",
"5 days". Unit can be of one "sec", "min", "hour", "day", "week",
"month", "year".}

\item{offset}{Use if you don't want breaks to start at zero, or on a
conventional date or time boundary such as the 1st of January or midnight.
Either a number, or for date/times, a single string of the form
\code{"{n} {unit}"}, as for \code{width}.

\code{offset} can be a vector, which will accumulate in the order given. This
is mostly useful for dates, where e.g. \code{c("3 months", "5 days")} will
offset by three months and five days, which is useful for the UK tax year.
Note that due to way that dates are rounded, there's no guarantee that
\code{offset = c(x, y)} will give the same result as \code{offset = c(y, x)}.}

\item{n}{number of breaks}
}
\description{
Generate minor breaks between major breaks either spaced with a fixed width,
or having a fixed number.
}
\examples{
\dontshow{if (requireNamespace("ggplot2", quietly = TRUE)) withAutoprint(\{ # examplesIf}
library(ggplot2)
ggplot(mpg, aes(x = displ, y = hwy)) +
  geom_point() +
  scale_x_continuous(minor_breaks = minor_breaks_width(1, offset = 0))
\dontshow{\}) # examplesIf}
demo_log10(c(1, 1e6))
demo_log10(c(1, 1e6), minor_breaks = minor_breaks_n(10))
}
