% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans-numeric.r
\name{modulus_trans}
\alias{modulus_trans}
\title{Modulus transformation.}
\usage{
modulus_trans(p)
}
\arguments{
\item{p}{Transformation exponent.}
}
\description{
Modulus transformation.
}
\details{
Implements a generalisation of the Box-Cox transformation that works
for data with both positive and negative values.
}
\references{
J. John and N. Draper. "An alternative family of
transformations." Applied Statistics, pages 190-197, 1980.
\url{http://www.jstor.org/stable/2986305}
}
