% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-auto.R
\name{number_auto}
\alias{number_auto}
\title{Automatic number formatter}
\usage{
number_auto(x)
}
\arguments{
\item{x}{A numeric vector to format.}
}
\description{
Switches between \code{\link[=number_format]{number_format()}} and \code{\link[=scientific_format]{scientific_format()}} based on a set of
heuristics designed to automatically generate useful labels across a wide
range of inputs
}
\examples{
# Very small and very large numbers get scientific notation
demo_continuous(c(0, 1e-6), labels = number_auto)
demo_continuous(c(0, 1e9), labels = number_auto)

# Other ranges get the numbers printed in full
demo_continuous(c(0, 1e-3), labels = number_auto)
demo_continuous(c(0, 1), labels = number_auto)
demo_continuous(c(0, 1e3), labels = number_auto)
demo_continuous(c(0, 1e6), labels = number_auto)
}
