% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-bytes.R
\name{number_bytes_format}
\alias{number_bytes_format}
\alias{number_bytes}
\title{Bytes formatter: convert to byte measurement and display symbol.}
\usage{
number_bytes_format(symbol = "auto", units = "binary", ...)

number_bytes(x, symbol = "auto", units = c("binary", "si"), ...)
}
\arguments{
\item{symbol}{byte symbol to use. If "auto" the symbol used will be
determined separately for each value of \code{x}. Valid symbols are "B", "kB",
"MB", "GB", "TB", "PB", "EB", "ZB", and "YB" for SI units, and the "iB"
variants for binary units.}

\item{units}{which unit base to use, "binary" (1024 base) or "si" (1000 base)}

\item{...}{other arguments passed to \code{\link[=number]{number()}}}

\item{x}{a numeric vector to format}
}
\value{
a function with three parameters, \code{x}, a numeric vector that returns
a character vector, \code{symbol} the byte symbol (e.g. "kB") desired and the
measurement \code{units} (traditional \code{binary} or \code{si} for SI metric units).
}
\description{
Bytes formatter: convert to byte measurement and display symbol.
}
\examples{
number_bytes_format()(sample(3000000000, 10))
number_bytes(sample(3000000000, 10))
number_bytes(sample(3000000000, 10), accuracy = .1)
number_bytes(1024^(0:4))
number_bytes(1024^(0:4), units = "si", accuracy = .01)
number_bytes(1000^(1:3), "kB", units = "si")
}
\references{
Units of Information (Wikipedia) :
\url{http://en.wikipedia.org/wiki/Units_of_information}
}
