% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offset-by.r
\name{offset_by}
\alias{offset_by}
\title{Add a fixed amount to numbers and dates}
\usage{
offset_by(x, size, ...)
}
\arguments{
\item{size}{interval size}

\item{...}{other arguments passed on to methods}
}
\description{
Add a fixed amount to numbers and dates
}
\seealso{
\code{\link[=fullseq]{fullseq()}}
}
\keyword{internal}
