% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.r
\name{oob_censor}
\alias{oob_censor}
\alias{censor}
\title{Censor any values outside of range}
\usage{
oob_censor(x, range = c(0, 1), only.finite = TRUE)

censor(x, range = c(0, 1), only.finite = TRUE)
}
\arguments{
\item{x}{numeric vector of values to manipulate.}

\item{range}{numeric vector of length two giving desired output range.}

\item{only.finite}{if \code{TRUE} (the default), will only modify
finite values.}
}
\description{
Censor any values outside of range
}
\examples{
oob_censor(c(-1, 0.5, 1, 2, NA))
}
