% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.r
\name{oob_discard}
\alias{oob_discard}
\alias{discard}
\title{Discard any values outside of range}
\usage{
oob_discard(x, range = c(0, 1))

discard(x, range = c(0, 1))
}
\arguments{
\item{x}{numeric vector of values to manipulate.}

\item{range}{numeric vector of length two giving desired output range.}
}
\description{
Discard any values outside of range
}
\examples{
oob_discard(c(-1, 0.5, 1, 2, NA))
}
\seealso{
Other out of bounds handlers: 
\code{\link{oob_censor}()},
\code{\link{oob_keep}()},
\code{\link{oob_squish_infinite}()},
\code{\link{oob_squish}()}
}
\concept{out of bounds handlers}
