% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.r
\name{oob_squish}
\alias{oob_squish}
\alias{squish}
\title{Squish values into range}
\usage{
oob_squish(x, range = c(0, 1), only.finite = TRUE)

squish(x, range = c(0, 1), only.finite = TRUE)
}
\arguments{
\item{x}{numeric vector of values to manipulate.}

\item{range}{numeric vector of length two giving desired output range.}

\item{only.finite}{if \code{TRUE} (the default), will only modify
finite values.}
}
\description{
Squish values into range
}
\examples{
oob_squish(c(-1, 0.5, 1, 2, NA))
oob_squish(c(-1, 0, 0.5, 1, 2))
}
\seealso{
Other out of bounds handlers: 
\code{\link{oob_censor}()},
\code{\link{oob_discard}()},
\code{\link{oob_keep}()},
\code{\link{oob_squish_infinite}()}
}
\author{
Homer Strong \href{mailto:homer.strong@gmail.com}{homer.strong@gmail.com}
}
\concept{out of bounds handlers}
