% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.r
\name{oob_squish_infinite}
\alias{oob_squish_infinite}
\alias{squish_infinite}
\title{Squish infinite values to range}
\usage{
oob_squish_infinite(x, range = c(0, 1))

squish_infinite(x, range = c(0, 1))
}
\arguments{
\item{x}{numeric vector of values to manipulate.}

\item{range}{numeric vector of length two giving desired output range.}
}
\description{
Squish infinite values to range
}
\examples{
oob_squish_infinite(c(-Inf, -1, 0, 1, 2, Inf))
}
\seealso{
Other out of bounds handlers: 
\code{\link{oob_censor}()},
\code{\link{oob_discard}()},
\code{\link{oob_keep}()},
\code{\link{oob_squish}()}
}
\concept{out of bounds handlers}
