% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatter.r
\name{ordinal_format}
\alias{ordinal_format}
\alias{ordinal}
\title{Ordinal formatter: add ordinal suffixes (-st, -nd, -rd, -th) to numbers.}
\usage{
ordinal_format(prefix = "", suffix = "", big.mark = ",", rules = NULL,
  ...)

ordinal(x, prefix = "", suffix = "", big.mark = ",", rules = NULL, ...)
}
\arguments{
\item{prefix, suffix}{Symbols to display before and after value}

\item{big.mark}{Character used between every 3 digits.}

\item{rules}{Custom rules for computing ordinal indicators}

\item{...}{Other arguments passed on to \code{\link{number_format}},
\code{NULL} for English (see examples).}

\item{x}{a numeric vector to format}
}
\value{
a function with single paramater x, a numeric vector, that
returns a character vector
}
\description{
Ordinal formatter: add ordinal suffixes (-st, -nd, -rd, -th) to numbers.
}
\examples{
ordinal_format()(1:10)
ordinal(1:10)

# Custom rules for French
french <- list(
  er = "^1$",
  nd = "^2$",
  e = "([0-9]+)[12]$",
  e = "[03456789]"
)
ordinal(1:20, rules = french)
}
