% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-area.R
\name{pal_area}
\alias{pal_area}
\alias{area_pal}
\alias{abs_area}
\title{Area palettes (continuous)}
\usage{
pal_area(range = c(1, 6))

area_pal(range = c(1, 6))

abs_area(max)
}
\arguments{
\item{range}{Numeric vector of length two, giving range of possible sizes.
Should be greater than 0.}

\item{max}{A number representing the maximum size.}
}
\description{
Area palettes (continuous)
}
