% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-dichromat.R
\name{pal_dichromat}
\alias{pal_dichromat}
\alias{dichromat_pal}
\title{Dichromat (colour-blind) palette (discrete)}
\usage{
pal_dichromat(name)

dichromat_pal(name)
}
\arguments{
\item{name}{Name of colour palette.  One of:
\Sexpr[results=rd,stage=build]{scales:::dichromat_schemes()}}
}
\description{
Dichromat (colour-blind) palette (discrete)
}
\examples{
if (requireNamespace("dichromat", quietly = TRUE)) {
  show_col(pal_dichromat("BluetoOrange.10")(10))
  show_col(pal_dichromat("BluetoOrange.10")(5))

  # Can use with gradient_n to create a continous gradient
  cols <- pal_dichromat("DarkRedtoBlue.12")(12)
  show_col(pal_gradient_n(cols)(seq(0, 1, length.out = 30)))
}
}
