% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-gradient.R
\name{pal_gradient_n}
\alias{pal_gradient_n}
\alias{gradient_n_pal}
\title{Arbitrary colour gradient palette (continuous)}
\usage{
pal_gradient_n(colours, values = NULL, space = "Lab")

gradient_n_pal(colours, values = NULL, space = "Lab")
}
\arguments{
\item{colours}{vector of colours}

\item{values}{if colours should not be evenly positioned along the gradient
this vector gives the position (between 0 and 1) for each colour in the
\code{colours} vector. See \code{\link[=rescale]{rescale()}} for a convenience function
to map an arbitrary range to between 0 and 1.}

\item{space}{colour space in which to calculate gradient. Must be "Lab" -
other values are deprecated.}
}
\description{
Arbitrary colour gradient palette (continuous)
}
