% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-gradient.R
\name{pal_gradient_seq}
\alias{pal_gradient_seq}
\title{Sequential colour gradient palette (continuous)}
\usage{
pal_gradient_seq(low = "#2B6788", high = "#90503F", space = "Lab")
}
\arguments{
\item{low}{colour for low end of gradient.}

\item{high}{colour for high end of gradient.}

\item{space}{colour space in which to calculate gradient. Must be "Lab" -
other values are deprecated.}
}
\description{
Sequential colour gradient palette (continuous)
}
\examples{
x <- seq(0, 1, length.out = 25)
show_col(pal_seq_gradient()(x))
show_col(pal_seq_gradient("white", "black")(x))

show_col(pal_seq_gradient("white", "#90503F")(x))
}
