% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-grey.R
\name{pal_grey}
\alias{pal_grey}
\alias{grey_pal}
\title{Grey scale palette (discrete)}
\usage{
pal_grey(start = 0.2, end = 0.8)

grey_pal(start = 0.2, end = 0.8)
}
\arguments{
\item{start}{grey value at low end of palette}

\item{end}{grey value at high end of palette}
}
\description{
Grey scale palette (discrete)
}
\examples{
show_col(pal_grey()(25))
show_col(pal_grey(0, 1)(25))
}
\seealso{
\code{\link[=pal_seq_gradient]{pal_seq_gradient()}} for continuous version
}
