% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-identity.R
\name{pal_identity}
\alias{pal_identity}
\alias{identity_pal}
\title{Identity palette}
\usage{
pal_identity()

identity_pal()
}
\description{
Leaves values unchanged - useful when the data is already scaled.
}
