% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-linetype.R
\name{pal_linetype}
\alias{pal_linetype}
\alias{linetype_pal}
\title{Line type palette (discrete)}
\usage{
pal_linetype()

linetype_pal()
}
\description{
Based on a set supplied by Richard Pearson, University of Manchester
}
