% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-manual.R
\name{pal_manual}
\alias{pal_manual}
\alias{manual_pal}
\title{Manual palette (discrete)}
\usage{
pal_manual(values)

manual_pal(values)
}
\arguments{
\item{values}{vector of values to be used as a palette.}
}
\description{
Manual palette (discrete)
}
