% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-seq.R
\name{pal_seq}
\alias{pal_seq}
\title{Sequence palette}
\usage{
pal_seq(min = 0, max = 1)
}
\arguments{
\item{min}{Scalar number setting the start.}

\item{max}{Scalar number setting the end.}
}
\value{
A discrete palette.
}
\description{
Creates a regular sequence between two numbers.
}
