% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-gradient.R
\name{pal_seq_gradient}
\alias{pal_seq_gradient}
\alias{seq_gradient_pal}
\title{Sequential colour gradient palette (continuous)}
\usage{
pal_seq_gradient(low = mnsl("10B 4/6"), high = mnsl("10R 4/6"), space = "Lab")

seq_gradient_pal(low = mnsl("10B 4/6"), high = mnsl("10R 4/6"), space = "Lab")
}
\arguments{
\item{low}{colour for low end of gradient.}

\item{high}{colour for high end of gradient.}

\item{space}{colour space in which to calculate gradient. Must be "Lab" -
other values are deprecated.}
}
\description{
Sequential colour gradient palette (continuous)
}
\examples{
x <- seq(0, 1, length.out = 25)
show_col(pal_seq_gradient()(x))
show_col(pal_seq_gradient("white", "black")(x))

library(munsell)
show_col(pal_seq_gradient("white", mnsl("10R 4/6"))(x))
}
