% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-shape.r
\name{pal_shape}
\alias{pal_shape}
\alias{shape_pal}
\title{Shape palette (discrete)}
\usage{
pal_shape(solid = TRUE)

shape_pal(solid = TRUE)
}
\arguments{
\item{solid}{should shapes be solid or not?}
}
\description{
Shape palette (discrete)
}
