% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatter.r
\name{parse_format}
\alias{parse_format}
\title{Parse a label to produce expressions for plotmath.}
\usage{
parse_format(text)
}
\arguments{
\item{text}{A character vector to parse. Other R objects will be coerced to
character if possible}
}
\value{
An expression object.
}
\description{
Parse a label to produce expressions for plotmath.
}
\examples{
parse_format(c("alpha", "beta", "gamma"))
parse_format(1:5)
}
\seealso{
\code{\link[=plotmath]{plotmath()}}
}
