% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.r
\name{pretty_breaks}
\alias{pretty_breaks}
\title{Pretty breaks}
\usage{
pretty_breaks(n = 5, ...)
}
\arguments{
\item{n}{desired number of breaks}

\item{...}{other arguments passed on to \code{\link[=pretty]{pretty()}}}
}
\description{
Uses default R break algorithm as implemented in \code{\link[=pretty]{pretty()}}. This is
primarily useful for date/times, as \code{\link[=extended_breaks]{extended_breaks()}} should do a slightly
better job for numeric scales.
}
\examples{
one_months <- as.POSIXct(c("2020-05-01", "2020-06-01"))
demo_datetime(two_months)
demo_datetime(two_months, breaks = pretty_breaks(10))
}
\keyword{internal}
