% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans-numeric.r
\name{pseudo_log_trans}
\alias{pseudo_log_trans}
\title{Pseudo-log transformation}
\usage{
pseudo_log_trans(sigma = 1, base = exp(1))
}
\arguments{
\item{sigma}{scaling factor for the linear part}

\item{base}{approximate logarithm base used}
}
\description{
A transformation mapping numbers to a signed logarithmic scale
with a smooth transition to linear scale around 0.
}
