% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans-numeric.r
\name{quantile_trans}
\alias{quantile_trans}
\alias{inv_logit_trans}
\alias{inv_probit_trans}
\title{Quantile transformation}
\usage{
quantile_trans(distribution, ...)

inv_logit_trans()

inv_probit_trans()
}
\arguments{
\item{distribution}{probability distribution.  Should be standard R
abbreviation so that "p" + distribution is a valid cumulative distribution
function, "q" + distribution is a valid quantile function, and
"d" + distribution is a valid probability density function.}

\item{...}{other arguments passed on to distribution, quantile, and density functions}
}
\description{
Quantile transformation
}
\examples{
plot(inv_logit_trans(), xlim = c(-3, 3))
plot(inv_probit_trans(), xlim = c(-3, 3))
}
\seealso{
\code{\link[=probability_trans]{probability_trans()}}, the inverse of this transformation
}
