% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.R
\name{Range}
\alias{Range}
\alias{DiscreteRange}
\alias{ContinuousRange}
\title{Mutable ranges}
\description{
Mutable ranges have a two methods (\code{train} and \code{reset}), and
make it possible to build up complete ranges with multiple passes.
}
