% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minor_breaks.R
\name{regular_minor_breaks}
\alias{regular_minor_breaks}
\title{Minor breaks}
\usage{
regular_minor_breaks(reverse = FALSE)
}
\arguments{
\item{reverse}{if TRUE, calculates the minor breaks for a reversed scale}
}
\description{
Places minor breaks between major breaks.
}
\examples{
m <- extended_breaks()(c(1, 10))
regular_minor_breaks()(m, c(1, 10), n = 2)

n <- extended_breaks()(c(0, -9))
regular_minor_breaks(reverse = TRUE)(n, c(0, -9), n = 2)
}
\keyword{internal}
