% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rescale_none}
\alias{rescale_none}
\title{Don't peform rescaling}
\usage{
rescale_none(x, ...)
}
\arguments{
\item{x}{numeric vector of values to manipulate.}

\item{...}{all other arguments ignored}
}
\description{
Don't peform rescaling
}
\examples{
rescale_none(1:100)
}

