% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans-numeric.R
\name{reverse_trans}
\alias{reverse_trans}
\title{Reverse transformation}
\usage{
reverse_trans()
}
\description{
reversing transformation works by multiplying the input with -1. This means
that reverse transformation cannot easily be composed with transformations
that require positive input unless the reversing is done as a final step.
}
\examples{
plot(reverse_trans(), xlim = c(-1, 1))
}
