% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatter.r
\name{scientific_format}
\alias{scientific_format}
\alias{scientific}
\title{Scientific formatter.}
\usage{
scientific_format(digits = 3, prefix = "", suffix = "", trim = TRUE,
  ...)

scientific(x, digits = 3, prefix = "", suffix = "", trim = TRUE, ...)
}
\arguments{
\item{digits}{number of significant digits to show}

\item{prefix, suffix}{symbols to display before and after value}

\item{trim}{logical, if \code{FALSE}, values are right-justified to a common
width (see \code{\link[base]{format}})}

\item{...}{other arguments passed on to \code{\link[base]{format}}}

\item{x}{a numeric vector to format}
}
\value{
a function with single parameter x, a numeric vector, that
returns a character vector
}
\description{
Scientific formatter.
}
\examples{
scientific_format()(1:10)
scientific_format()(runif(10))
scientific_format(digits = 2)(runif(10))
scientific(1:10)
scientific(runif(10))
scientific(runif(10), digits = 2)
}
