% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colour-manip.r
\name{show_col}
\alias{show_col}
\title{Show colours.}
\usage{
show_col(colours)
}
\arguments{
\item{colours}{a character vector of colours}
}
\description{
A quick and dirty way to show colours in a plot.
}

