% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show_linetype}
\alias{show_linetype}
\title{Show linetypes}
\usage{
show_linetype(linetypes, labels = TRUE)
}
\arguments{
\item{linetypes}{A character vector of linetypes. See \code{\link[=par]{par()}} for
linetype values.}

\item{labels}{Label each line with its linetype (\code{lty}) value.}
}
\value{
This function called for the side effect of creating a plot.
It returns \code{linetypes}.
}
\description{
A quick and dirty way to show linetypes in a plot.
}
\seealso{
Other `show_*` functions: \code{\link{show_col}},
  \code{\link{show_shape}}
}
