% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show_shape}
\alias{show_shape}
\title{Show shapes.}
\usage{
show_shape(shapes, labels = TRUE)
}
\arguments{
\item{shapes}{A numeric or character vector of shapes. See
\code{\link[=par]{par()}}.}

\item{labels}{Include the plotting character value of the symbol.}
}
\value{
This function is called for the side effect of creating a plot, but
it invisibly returns \code{shapes}.
}
\description{
A quick and dirty way to show shapes in a plot.
}
\seealso{
Other `show_*` functions: \code{\link{show_col}},
  \code{\link{show_linetype}}
}
