% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{si_prefixes}
\alias{si_prefixes}
\title{Metric prefixes used in the SI system}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 21 rows and 3 columns.}
\usage{
si_prefixes
}
\description{
Names, symbols and powers of metric prefixes in the SI system. Taken from
\url{https://en.wikipedia.org/wiki/Metric_prefix}.
}
\keyword{datasets}
