% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si-powers.R
\name{si_scale}
\alias{si_scale}
\alias{si_prefix}
\title{SI scale factors and unit prefixes}
\usage{
si_scale(x, base = 10)

si_prefix(x, base = 10)
}
\arguments{
\item{x}{A numeric vector.}

\item{base}{A numeric vector of bases whose SI powers are included.}
}
\value{
\code{si_scale()}: A numeric vector of SI scaling factors.

\code{si_prefix()}: A character vector of SI unit prefixes.
}
\description{
SI scale factors and unit prefixes
}
\examples{
# scaling factors
si_scale(10^(-6:6))
si_scale(10^(-6:6), base = 1000)

# unit prefixes
si_prefix(10^(-6:6))
si_prefix(10^(-6:6), base = 1000)
si_prefix(10^(-6:6), base = c(100, 1000))
}
\seealso{
\code{\link[=number]{number()}} for formatting numbers
}
