% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans-numeric.r
\name{sqrt_trans}
\alias{sqrt_trans}
\title{Square-root transformation}
\usage{
sqrt_trans()
}
\description{
This is the variance stabilising transformation for the Poisson
distribution.
}
\examples{
plot(sqrt_trans(), xlim = c(0, 5))
}
