% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-continuous.R
\name{train_continuous}
\alias{train_continuous}
\title{Train (update) a continuous scale}
\usage{
train_continuous(new, existing = NULL, call = caller_env())
}
\arguments{
\item{new}{New data to add to scale}

\item{existing}{Optional existing scale to update}

\item{call}{A call to display in error messages}
}
\description{
Strips attributes and always returns a numeric vector
}
