% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans-numeric.r
\name{trans_atanh}
\alias{trans_atanh}
\alias{atanh_trans}
\title{Arc-tangent transformation}
\usage{
trans_atanh()

atanh_trans()
}
\description{
Arc-tangent transformation
}
\examples{
plot(atanh_trans(), xlim = c(-1, 1))
}
