% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans-numeric.r
\name{trans_exp}
\alias{trans_exp}
\alias{exp_trans}
\alias{identity_trans}
\title{Exponential transformation (inverse of log transformation)}
\usage{
trans_exp(base = exp(1))

exp_trans(base = exp(1))

identity_trans()
}
\arguments{
\item{base}{Base of logarithm}
}
\description{
Exponential transformation (inverse of log transformation)
}
\examples{
plot(trans_exp(0.5), xlim = c(-2, 2))
plot(trans_exp(1), xlim = c(-2, 2))
plot(trans_exp(2), xlim = c(-2, 2))
plot(trans_exp(), xlim = c(-2, 2))
}
