% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans-numeric.r
\name{trans_identity}
\alias{trans_identity}
\title{Identity transformation (do nothing)}
\usage{
trans_identity()
}
\description{
Identity transformation (do nothing)
}
\examples{
plot(trans_identity(), xlim = c(-1, 1))
}
