% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans-numeric.r
\name{trans_log}
\alias{trans_log}
\alias{log_trans}
\alias{trans_log10}
\alias{log10_trans}
\alias{trans_log2}
\alias{log2_trans}
\alias{trans_log1p}
\alias{log1p_trans}
\alias{trans_pseudo_log}
\alias{pseudo_log_trans}
\title{Log transformations}
\usage{
trans_log(base = exp(1))

log_trans(base = exp(1))

trans_log10()

log10_trans()

trans_log2()

log2_trans()

trans_log1p()

log1p_trans()

trans_pseudo_log(sigma = 1, base = exp(1))

pseudo_log_trans(sigma = 1, base = exp(1))
}
\arguments{
\item{base}{base of logarithm}

\item{sigma}{Scaling factor for the linear part of pseudo-log transformation.}
}
\description{
\itemize{
\item \code{trans_log()}: \code{log(x)}
\item \code{trans_log1p()}: \code{log(x + 1)}
\item \code{trans_pseudo_log()}:  smoothly transition to linear scale around 0.
}
}
\examples{
plot(trans_log2(), xlim = c(0, 5))
plot(trans_log(), xlim = c(0, 5))
plot(trans_log10(), xlim = c(0, 5))

plot(trans_log(), xlim = c(0, 2))
plot(trans_log1p(), xlim = c(-1, 1))

# The pseudo-log is defined for all real numbers
plot(trans_pseudo_log(), xlim = c(-5, 5))
lines(trans_log(), xlim = c(0, 5), col = "red")

# For large positives nubmers it's very close to log
plot(trans_pseudo_log(), xlim = c(1, 20))
lines(trans_log(), xlim = c(1, 20), col = "red")
}
