% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans-numeric.r
\name{trans_probability}
\alias{trans_probability}
\alias{probability_trans}
\alias{trans_logit}
\alias{logit_trans}
\alias{trans_probit}
\alias{probit_trans}
\title{Probability transformation}
\usage{
trans_probability(distribution, ...)

probability_trans(distribution, ...)

trans_logit()

logit_trans()

trans_probit()

probit_trans()
}
\arguments{
\item{distribution}{probability distribution.  Should be standard R
abbreviation so that "p" + distribution is a valid probability density
function, and "q" + distribution is a valid quantile function.}

\item{...}{other arguments passed on to distribution and quantile functions}
}
\description{
Probability transformation
}
\examples{
plot(trans_logit(), xlim = c(0, 1))
plot(trans_probit(), xlim = c(0, 1))
}
