% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans-numeric.r
\name{trans_reciprocal}
\alias{trans_reciprocal}
\alias{reciprocal_trans}
\title{Reciprocal transformation}
\usage{
trans_reciprocal()

reciprocal_trans()
}
\description{
Reciprocal transformation
}
\examples{
plot(trans_reciprocal(), xlim = c(0, 1))
}
