% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans-numeric.r
\name{trans_reverse}
\alias{trans_reverse}
\alias{reverse_trans}
\title{Reverse transformation}
\usage{
trans_reverse()

reverse_trans()
}
\description{
Reverse transformation
}
\examples{
plot(trans_reverse(), xlim = c(-1, 1))
}
