% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans-numeric.r
\name{trans_sqrt}
\alias{trans_sqrt}
\alias{sqrt_trans}
\title{Square-root transformation}
\usage{
trans_sqrt()

sqrt_trans()
}
\description{
This is the variance stabilising transformation for the Poisson
distribution.
}
\examples{
plot(trans_sqrt(), xlim = c(0, 5))
}
