% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-numeric.R
\name{transform_asinh}
\alias{transform_asinh}
\alias{asinh_trans}
\title{Inverse Hyperbolic Sine transformation}
\usage{
transform_asinh()

asinh_trans()
}
\description{
Inverse Hyperbolic Sine transformation
}
\examples{
plot(transform_asinh(), xlim = c(-1e2, 1e2))
}
