% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-numeric.R
\name{transform_asn}
\alias{transform_asn}
\alias{asn_trans}
\title{Arc-sin square root transformation}
\usage{
transform_asn()

asn_trans()
}
\description{
This is the variance stabilising transformation for the binomial
distribution.
}
\examples{
plot(transform_asn(), xlim = c(0, 1))
}
