% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-numeric.R
\name{transform_atanh}
\alias{transform_atanh}
\alias{atanh_trans}
\title{Arc-tangent transformation}
\usage{
transform_atanh()

atanh_trans()
}
\description{
Arc-tangent transformation
}
\examples{
plot(transform_atanh(), xlim = c(-1, 1))
}
