% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-compose.R
\name{transform_compose}
\alias{transform_compose}
\alias{compose_trans}
\title{Compose two or more transformations together}
\usage{
transform_compose(...)

compose_trans(...)
}
\arguments{
\item{...}{One or more transformers, either specified with string or
as individual transformer objects.}
}
\description{
This transformer provides a general mechanism for composing two or more
transformers together. The most important use case is to combine reverse
with other transformations.
}
\examples{
demo_continuous(10^c(-2:4), trans = "log10", labels = label_log())
demo_continuous(10^c(-2:4), trans = c("log10", "reverse"), labels = label_log())
}
