% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-date.R
\name{transform_date}
\alias{transform_date}
\alias{date_trans}
\title{Transformation for dates (class Date)}
\usage{
transform_date()

date_trans()
}
\description{
Transformation for dates (class Date)
}
\examples{
years <- seq(as.Date("1910/1/1"), as.Date("1999/1/1"), "years")
t <- transform_date()
t$transform(years)
t$inverse(t$transform(years))
t$format(t$breaks(range(years)))
}
