% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-numeric.R
\name{transform_identity}
\alias{transform_identity}
\alias{identity_trans}
\title{Identity transformation (do nothing)}
\usage{
transform_identity()

identity_trans()
}
\description{
Identity transformation (do nothing)
}
\examples{
plot(transform_identity(), xlim = c(-1, 1))
}
