% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-numeric.R
\name{transform_probability}
\alias{transform_probability}
\alias{transform_logit}
\alias{transform_probit}
\alias{probability_trans}
\alias{logit_trans}
\alias{probit_trans}
\title{Probability transformation}
\usage{
transform_probability(distribution, ...)

transform_logit()

transform_probit()

probability_trans(distribution, ...)

logit_trans()

probit_trans()
}
\arguments{
\item{distribution}{probability distribution.  Should be standard R
abbreviation so that "p" + distribution is a valid cumulative distribution
function, "q" + distribution is a valid quantile function, and
"d" + distribution is a valid probability density function.}

\item{...}{other arguments passed on to distribution and quantile functions}
}
\description{
Probability transformation
}
\examples{
plot(transform_logit(), xlim = c(0, 1))
plot(transform_probit(), xlim = c(0, 1))
}
