% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-numeric.R
\name{transform_reciprocal}
\alias{transform_reciprocal}
\alias{reciprocal_trans}
\title{Reciprocal transformation}
\usage{
transform_reciprocal()

reciprocal_trans()
}
\description{
Reciprocal transformation
}
\examples{
plot(transform_reciprocal(), xlim = c(0, 1))
}
