% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-numeric.R
\name{transform_sqrt}
\alias{transform_sqrt}
\alias{sqrt_trans}
\title{Square-root transformation}
\usage{
transform_sqrt()

sqrt_trans()
}
\description{
This is the variance stabilising transformation for the Poisson
distribution.
}
\examples{
plot(transform_sqrt(), xlim = c(0, 5))
}
