% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans-numeric.r
\name{transform_with}
\alias{transform_with}
\title{Transform vector with a function}
\usage{
transform_with(x, i, f)
}
\arguments{
\item{x}{vector}

\item{i}{logical vector specifying condition (ex. \code{x >= 0})}

\item{f}{function that does the transform}
}
\description{
Transforms x at locations i with function f. If i is NA, it will not
transform the value.
}
\keyword{internal}
