% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{trim_to_domain}
\alias{trim_to_domain}
\alias{trans_range}
\title{Compute range of transformed values}
\usage{
trim_to_domain(transform, x)

trans_range(transform, x)
}
\arguments{
\item{transform}{a transformation object, or the name of a transformation object
given as a string.}

\item{x}{a numeric vector to compute the range of}
}
\description{
Silently drops any ranges outside of the domain of \code{transform}.
}
\keyword{internal}
