% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-wrap.R
\name{wrap_format}
\alias{wrap_format}
\title{Wrap labels to fixed width}
\usage{
wrap_format(width)
}
\arguments{
\item{width}{value above which to wrap}
}
\value{
Function with single parameter x, a character vector, that
returns a wrapped character vector
}
\description{
Wrap labels to fixed width
}
\examples{
wrap_10 <- wrap_format(10)
wrap_10('A long line that needs to be wrapped')
}
