% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-area.r
\name{abs_area}
\alias{abs_area}
\title{Point area palette (continuous), with area proportional to value.}
\usage{
abs_area(max)
}
\arguments{
\item{max}{A number representing the maximum size.}
}
\description{
Point area palette (continuous), with area proportional to value.
}
