% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour-manip.r
\name{alpha}
\alias{alpha}
\title{Modify colour transparency}
\usage{
alpha(colour, alpha = NA, color_fmt = "character")
}
\arguments{
\item{colour}{colour}

\item{alpha}{new alpha level in [0,1].  If alpha is \code{NA},
existing alpha values are preserved.}

\item{color_fmt}{A string, whether the output should be a character vector
with colours (\code{"character"}) or an integer vector in the native color
format (\code{"native"})}
}
\description{
Vectorised in both colour and alpha.
}
\examples{
alpha("red", 0.1)
alpha(colours(), 0.5)
alpha("red", seq(0, 1, length.out = 10))
alpha(c("first" = "gold", "second" = "lightgray", "third" = "#cd7f32"), .5)
}
