% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour-manip.R
\name{alpha}
\alias{alpha}
\title{Modify colour transparency}
\usage{
alpha(colour, alpha = NA)
}
\arguments{
\item{colour}{colour}

\item{alpha}{new alpha level in [0,1].  If alpha is \code{NA},
existing alpha values are preserved.}
}
\description{
Vectorised in both colour and alpha.
}
\examples{
alpha("red", 0.1)
alpha(colours(), 0.5)
alpha("red", seq(0, 1, length.out = 10))
alpha(c("first" = "gold", "second" = "lightgray", "third" = "#cd7f32"), .5)
}
\seealso{
Other colour manipulation: 
\code{\link{col2hcl}()},
\code{\link{col_mix}()},
\code{\link{colour_manip}},
\code{\link{muted}()}
}
\concept{colour manipulation}
