% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colour-manip.r
\name{alpha}
\alias{alpha}
\title{Modify colour transparency.
Vectorised in both colour and alpha.}
\usage{
alpha(colour, alpha = NA)
}
\arguments{
\item{colour}{colour}

\item{alpha}{new alpha level in [0,1].  If alpha is \code{NA},
existing alpha values are preserved.}
}
\description{
Modify colour transparency.
Vectorised in both colour and alpha.
}
\examples{
alpha("red", 0.1)
alpha(colours(), 0.5)
alpha("red", seq(0, 1, length = 10))
}

