% Generated by roxygen2 (4.0.0): do not edit by hand
\name{area_pal}
\alias{area_pal}
\title{Point area palette (continuous).}
\usage{
area_pal(range = c(1, 6))
}
\arguments{
\item{range}{Numeric vector of length two, giving range of possible sizes.
Should be greater than 0.}
}
\description{
Point area palette (continuous).
}

