% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans-numeric.r
\name{boxcox_trans}
\alias{boxcox_trans}
\title{Box-Cox (Modulus) power transformation.}
\usage{
boxcox_trans(p)
}
\arguments{
\item{p}{Exponent of box-cox transformation.}
}
\description{
Box-Cox (Modulus) power transformation.
}
\details{
Implements a generalisation of the box-cox transformation that works
for values with both positive and negative values.
}
\references{
J. John and N. Draper. "An alternative family of
transformations." Applied Statistics, pages 190-197, 1980.
\url{http://www.jstor.org/stable/2986305}
}
