% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans-numeric.r
\name{boxcox_trans}
\alias{boxcox_trans}
\title{Box-Cox power transformation.}
\usage{
boxcox_trans(p)
}
\arguments{
\item{p}{Exponent of Box-Cox transformation.}
}
\description{
Box-Cox power transformation.
}
\details{
The Box-Cox transformation is suitable for strictly positive values.
}
\references{
Box, G. E., & Cox, D. R. (1964). An analysis of transformations.
Journal of the Royal Statistical Society. Series B (Methodological), 211-252.
\url{https://www.jstor.org/stable/2984418}
}
