% Generated by roxygen2 (4.0.0): do not edit by hand
\name{boxcox_trans}
\alias{boxcox_trans}
\title{Box-Cox power transformation.}
\usage{
boxcox_trans(p)
}
\arguments{
\item{p}{Exponent of boxcox transformation.}
}
\description{
Box-Cox power transformation.
}
\references{
See \url{http://en.wikipedia.org/wiki/Power_transform} for
}

