% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{breaks_exp}
\alias{breaks_exp}
\title{Breaks for exponentially transformed data}
\usage{
breaks_exp(n = 5, ...)
}
\arguments{
\item{n}{Desired number of breaks. You may get slightly more or fewer
breaks that requested.}

\item{...}{other arguments passed on to \code{\link[labeling:extended]{labeling::extended()}}}
}
\description{
This breaks function typically labels zero and the last \code{n - 1} integers of a
range if that range is large enough (currently: 3). For smaller ranges, it
uses \code{\link[=breaks_extended]{breaks_extended()}}.
}
\examples{
# Small range
demo_continuous(c(100, 102), transform = "exp", breaks = breaks_exp())
# Large range
demo_continuous(c(0, 100), transform = "exp", breaks = breaks_exp(n = 4))
}
