% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.r
\name{breaks_extended}
\alias{breaks_extended}
\alias{extended_breaks}
\title{Extended breaks}
\usage{
breaks_extended(n = 5, ...)
}
\arguments{
\item{n}{desired number of breaks}

\item{...}{other arguments passed on to \code{\link[labeling:extended]{labeling::extended()}}}
}
\description{
Uses Wilkinson's extended breaks algorithm as implemented in the
\pkg{labeling} package.
}
\examples{
breaks_extended()(1:10)
breaks_extended()(1:100)
}
\references{
Talbot, J., Lin, S., Hanrahan, P. (2010) An Extension of
Wilkinson's Algorithm for Positioning Tick Labels on Axes, InfoVis
2010.
}
