% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.r
\name{breaks_width}
\alias{breaks_width}
\title{Produce breaks spaced by a fixed distance}
\usage{
breaks_width(width, offset = 0)
}
\arguments{
\item{width}{Distance between each break}

\item{offset}{Use if you don't want breaks to start at zero}
}
\description{
Produce breaks spaced by a fixed distance
}
\examples{
demo_continuous(c(0, 100))
demo_continuous(c(0, 100), breaks = breaks_width(10))
demo_continuous(c(0, 100), breaks = breaks_width(20, -4))
demo_continuous(c(0, 100), breaks = breaks_width(20, 4))
}
