% Generated by roxygen2 (4.0.0): do not edit by hand
\name{col2hcl}
\alias{col2hcl}
\title{Modify standard R colour in hcl colour space.}
\usage{
col2hcl(colour, h, c, l, alpha = 1)
}
\arguments{
\item{colour}{character vector of colours to be modified}

\item{h}{new hue}

\item{l}{new luminance}

\item{c}{new chroma}

\item{alpha}{alpha value.  Defaults to 1.}
}
\description{
Transforms rgb to hcl, sets non-missing arguments and then backtransforms
to rgb.
}
\examples{
col2hcl(colors())
}

