\name{comma_format}
\alias{comma_format}
\title{Comma formatter: format number with commas separating thousands.}
\usage{comma_format(...)}

\description{
  Comma formatter: format number with commas separating
  thousands.
}
\value{a function with single paramater x, a numeric vector, that
returns a character vector}
\arguments{
  \item{...}{other arguments passed on to \code{\link{format}}}
}
