% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatter.r
\name{comma_format}
\alias{comma_format}
\alias{comma}
\title{Comma formatter: format number with commas separating thousands.}
\usage{
comma_format(...)

comma(x, ...)
}
\arguments{
\item{...}{other arguments passed on to \code{\link{format}}}

\item{x}{a numeric vector to format}
}
\value{
a function with single parameter x, a numeric vector, that
  returns a character vector
}
\description{
Comma formatter: format number with commas separating thousands.
}
\examples{
comma_format()(c(1, 1e3, 2000, 1e6))
comma_format(digits = 9)(c(1, 1e3, 2000, 1e6))
comma(c(1, 1e3, 2000, 1e6))

# If you're European you can switch . and , with the more general
# format_format
point <- format_format(big.mark = ".", decimal.mark = ",", scientific = FALSE)
point(c(1, 1e3, 2000, 1e6))
point(c(1, 1.021, 1000.01))
}
