% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans-date.r
\name{date_breaks}
\alias{date_breaks}
\title{Regularly spaced dates.}
\usage{
date_breaks(width = "1 month")
}
\arguments{
\item{width}{an interval specification, one of "sec", "min", "hour",
"day", "week", "month", "year". Can be by an integer and a space, or
followed by "s".}
}
\description{
Regularly spaced dates.
}
