% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks-retired.R
\name{date_breaks}
\alias{date_breaks}
\title{Regularly spaced dates}
\usage{
date_breaks(width = "1 month")
}
\arguments{
\item{width}{an interval specification, one of "sec", "min", "hour",
"day", "week", "month", "year". Can be by an integer and a space, or
followed by "s". Fractional seconds are supported.}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("retired")}
Use \code{breaks_width()} instead.
}
\keyword{internal}
