% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-date.R
\name{date_format}
\alias{date_format}
\alias{date_short}
\alias{time_format}
\title{Formatted dates and times}
\usage{
date_format(format = "\%Y-\%m-\%d", tz = "UTC")

date_short(format = c("\%Y", "\%b", "\%d", "\%H:\%M"), sep = "\\n")

time_format(format = "\%H:\%M:\%S", tz = "UTC")
}
\arguments{
\item{format}{For \code{date_format()} and \code{time_format()} a date/time format
string using standard POSIX specification.  See \code{\link[=strptime]{strptime()}} for details.

For \code{date_short()} a character vector of length 4 giving the format
components to use for year, month, day, and hour respectively.}

\item{tz}{a time zone name, see \code{\link[=timezones]{timezones()}}. Defaults
to UTC}

\item{sep}{Separator to use when combining date formats into a single string.}
}
\description{
\code{date_short()} attempts to construct the shortest format string
that will uniquely identify dates. It's inspired by matplotlib's
\href{https://matplotlib.org/api/dates_api.html#matplotlib.dates.ConciseDateFormatter}{ConciseDateFormatter},
but uses a slightly different approach: \code{ConciseDateFormatter} formats
"firsts" (e.g. first day of month, first day of day) specially;
\code{date_short()} formats changes (e.g. new month, new year) specially.
}
\examples{
date_range <- function(start, days) {
  start <- as.POSIXct(start)
  c(start, start + days * 24 * 60 * 60)
}

two_months <- date_range("2020-05-01", 60)
demo_datetime(two_months)
demo_datetime(two_months, labels = date_format("\%m/\%d"))
# ggplot2 provides a short-hand:
demo_datetime(two_months, date_labels = "\%m/\%d")

# An alternative labelling system is date_short()
demo_datetime(two_months, date_breaks = "7 days", labels = date_short())
# This is particularly effective for dense labels
one_year <- date_range("2020-05-01", 365)
demo_datetime(one_year, date_breaks = "month")
demo_datetime(one_year, date_breaks = "month", labels = date_short())
}
