% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-date.R
\name{date_format}
\alias{date_format}
\alias{time_format}
\title{Superseded interface to \code{label_date()}/\code{label_time()}}
\usage{
date_format(format = "\%Y-\%m-\%d", tz = "UTC", locale = NULL)

time_format(format = "\%H:\%M:\%S", tz = "UTC", locale = NULL)
}
\arguments{
\item{format}{For \code{label_date()} and \code{label_time()} a date/time format
string using standard POSIX specification.  See \code{\link[=strptime]{strptime()}} for details.

For \code{label_date_short()} a character vector of length 4 giving the format
components to use for year, month, day, and hour respectively.}

\item{tz}{a time zone name, see \code{\link[=timezones]{timezones()}}. Defaults
to UTC}

\item{locale}{Locale to use when for day and month names. The default
uses the current locale. Setting this argument requires stringi, and you
can see a complete list of supported locales with
\code{\link[stringi:stri_locale_list]{stringi::stri_locale_list()}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

These functions are kept for backward compatibility; you should switch
to \code{\link[=label_date]{label_date()}}/\code{\link[=label_time]{label_time()}} for new code.
}
\keyword{internal}
