% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans-date.r
\name{date_trans}
\alias{date_trans}
\title{Transformation for dates (class Date)}
\usage{
date_trans()
}
\description{
Transformation for dates (class Date)
}
\examples{
years <- seq(as.Date("1910/1/1"), as.Date("1999/1/1"), "years")
t <- date_trans()
t$transform(years)
t$inverse(t$transform(years))
t$format(t$breaks(range(years)))
}
