% Generated by roxygen2 (4.0.0): do not edit by hand
\name{discard}
\alias{discard}
\title{Discard any values outside of range.}
\usage{
discard(x, range = c(0, 1))
}
\arguments{
\item{x}{numeric vector of values to manipulate.}

\item{range}{numeric vector of length two giving desired output range.}
}
\description{
Discard any values outside of range.
}
\examples{
discard(c(-1, 0.5, 1, 2, NA))
}

