\name{dollar_format}
\alias{dollar_format}
\title{Currency formatter: round to nearest cent and display dollar sign.}

\description{
  Currency formatter: round to nearest cent and display
  dollar sign.
}
\value{a function with single paramater x, a numeric vector, that
returns a character vector}
