\name{format_format}
\alias{format_format}
\title{Format with using any arguments to \link{format}.}
\usage{format_format(...)}

\description{
  Format with using any arguments to \code{\link{format}}.
}
\seealso{\code{\link{format}}, \code{\link{format.Date}},
\code{\link{format.POSIXct}}}
\arguments{
  \item{...}{other arguments passed on to \code{\link{format}}.}
}
