% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels-retired.R
\name{format_format}
\alias{format_format}
\title{Label using \code{format()}}
\usage{
format_format(...)
}
\arguments{
\item{...}{Arguments passed on to \code{\link[=format]{format()}}.}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("retired")}
This function is kept for backward compatiblity; you should either use
\code{\link[=label_number]{label_number()}} or \code{\link[=label_date]{label_date()}} instead.
}
\keyword{internal}
