% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatter.r
\name{format_format}
\alias{format_format}
\title{Format with using any arguments to \code{\link{format}}.}
\usage{
format_format(...)
}
\arguments{
\item{...}{other arguments passed on to \code{\link{format}}.}
}
\description{
If the breaks have names, they will be used in preference to formatting
the breaks.
}
\seealso{
\code{\link{format}}, \code{\link{format.Date}},
  \code{\link{format.POSIXct}}
}

