% Generated by roxygen2 (4.0.0): do not edit by hand
\name{gradient_n_pal}
\alias{gradient_n_pal}
\title{Arbitrary colour gradient palette (continous).}
\usage{
gradient_n_pal(colours, values = NULL, space = "Lab")
}
\arguments{
\item{colours}{vector of colours}

\item{values}{if colours should not be evenly positioned along the gradient
this vector gives the position (between 0 and 1) for each colour in the
\code{colours} vector. See \code{\link{rescale}} for a convience function
to map an arbitrary range to between 0 and 1.}

\item{space}{colour space in which to calculate gradient.  "Lab" usually
best unless gradient goes through white.}
}
\description{
Arbitrary colour gradient palette (continous).
}

