% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-grey.r
\name{grey_pal}
\alias{grey_pal}
\title{Grey scale palette (discrete)}
\usage{
grey_pal(start = 0.2, end = 0.8)
}
\arguments{
\item{start}{grey value at low end of palette}

\item{end}{grey value at high end of palette}
}
\description{
Grey scale palette (discrete)
}
\examples{
show_col(grey_pal()(25))
show_col(grey_pal(0, 1)(25))
}
\seealso{
\code{\link[=seq_gradient_pal]{seq_gradient_pal()}} for continuous version
}
