% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-bytes.R
\name{label_bytes}
\alias{label_bytes}
\alias{number_bytes_format}
\alias{number_bytes}
\title{Bytes formatter: convert to byte measurement and display symbol}
\usage{
label_bytes(symbol = "auto", units = c("si", "binary"), ...)

number_bytes_format(symbol = "auto", units = "binary", ...)

number_bytes(x, symbol = "auto", units = c("binary", "si"),
  accuracy = 1, ...)
}
\arguments{
\item{symbol}{byte symbol to use. If "auto" the symbol used will be
determined separately for each value of \code{x}. Valid symbols are "B", "kB",
"MB", "GB", "TB", "PB", "EB", "ZB", and "YB" for SI units, and the "iB"
variants for binary units.}

\item{units}{which unit base to use, "binary" (1024 base) or "si" (1000 base)}

\item{...}{other arguments passed to \code{\link[=number]{number()}}}

\item{x}{A numeric vector to format.}

\item{accuracy}{Number to round to. Use (e.g.) \code{0.01} to show 2 decimal
places of precision. If \code{NULL}, the default, uses a heuristic that should
ensure breaks have the minimum number of digits needed to show the
difference between adjacent values.

Applied to rescaled data.}
}
\value{
a function with three parameters, \code{x}, a numeric vector that returns
a character vector, \code{symbol} the byte symbol (e.g. "kB") desired and the
measurement \code{units} (traditional \code{binary} or \code{si} for SI metric units).
}
\description{
Bytes formatter: convert to byte measurement and display symbol
}
\examples{
demo_continuous(c(1, 1e6))
demo_continuous(c(1, 1e6), label = label_bytes())
# Force all to use same units
demo_continuous(c(1, 1e6), label = label_bytes("kB"))

# Auto units are particularly nice on log scales
demo_log10(c(1, 1e6))
demo_log10(c(1, 1e6), label = label_bytes())

# You can also use binary units where a megabyte is defined as
# (1024) ^ 2 bytes rather than (1000) ^ 2. You'll need to override
# the default breaks to make this more informative.
demo_continuous(c(1, 1e6), label = label_bytes(units = "binary"))
demo_continuous(c(1, 1e6),
  breaks = breaks_width(250 * 1024),
  label = label_bytes(units = "binary")
)
}
\references{
Units of Information (Wikipedia) :
\url{http://en.wikipedia.org/wiki/Units_of_information}
}
