% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-number-si.R
\name{label_number_si}
\alias{label_number_si}
\title{Label numbers, rescaled automatically to use SI prefixes}
\usage{
label_number_si(accuracy = 1, prefix = "", suffix = "", ...)
}
\arguments{
\item{accuracy}{Number to round to. Use (e.g.) \code{0.01} to show 2 decimal
places of precision. If \code{NULL}, the default, uses a heuristic that should
ensure breaks have the minimum number of digits needed to show the
difference between adjacent values.

Applied to rescaled data.}

\item{prefix}{Symbols to display before and after value.}

\item{suffix}{Symbols to display before and after value.}

\item{...}{Other arguments passed on to \code{\link[base:format]{base::format()}}.}
}
\description{
\code{number_si()} automatically scales and labels with the best SI prefix,
"K" for values \eqn{\ge} 10e3, "M" for \eqn{\ge} 10e6,
"B" for \eqn{\ge} 10e9, and "T" for \eqn{\ge} 10e12.
}
\examples{
demo_continuous(c(1, 1e9), label = label_number_si())
demo_log10(c(1, 1e9), breaks = log_breaks(10), labels = label_number_si())
}
