% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-percent.R
\name{label_percent}
\alias{label_percent}
\alias{percent_format}
\alias{percent}
\title{Label percentages (2.5\%, 50\%, etc)}
\usage{
label_percent(accuracy = NULL, scale = 100, prefix = "",
  suffix = "\%", big.mark = " ", decimal.mark = ".", trim = TRUE,
  ...)

percent_format(accuracy = NULL, scale = 100, prefix = "",
  suffix = "\%", big.mark = " ", decimal.mark = ".", trim = TRUE,
  ...)

percent(x, accuracy = NULL, scale = 100, prefix = "",
  suffix = "\%", big.mark = " ", decimal.mark = ".", trim = TRUE,
  ...)
}
\arguments{
\item{accuracy}{Number to round to. Use (e.g.) \code{0.01} to show 2 decimal
places of precision. If \code{NULL}, the default, uses a heuristic that should
ensure breaks have the minimum number of digits needed to show the
difference between adjacent values.

Applied to rescaled data.}

\item{scale}{A scaling factor: \code{x} will be multiply by \code{scale} before
formating (useful if the underlying data is on another scale,
e.g. for computing percentages or thousands).}

\item{prefix}{Symbols to display before and after value.}

\item{suffix}{Symbols to display before and after value.}

\item{big.mark}{Character used between every 3 digits to separate thousands.}

\item{decimal.mark}{The character to be used to indicate the numeric
decimal point.}

\item{trim}{Logical, if \code{FALSE}, values are right-justified to a common
width (see \code{\link[base:format]{base::format()}}).}

\item{...}{Other arguments passed on to \code{\link[base:format]{base::format()}}.}

\item{x}{A numeric vector to format.}
}
\value{
\code{*_format()} returns a function with single parameter
\code{x}, a numeric vector, that returns a character vector.
}
\description{
Label percentages (2.5\%, 50\%, etc)
}
\section{Old interface}{

\code{percent()} and \code{percent_format()} are retired; please use \code{label_percent()}
instead.
}

\examples{
demo_continuous(c(0, 1))
demo_continuous(c(0, 1), labels = label_percent())

# Use prefix and suffix to create your own variants
french_percent <- label_percent(
  decimal.mark = ",",
  suffix = " \%"
)
demo_continuous(c(0, .01), labels = french_percent)
}
