% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-pvalue.R
\name{label_pvalue}
\alias{label_pvalue}
\title{Label p-values (e.g. <0.001, 0.25, p >= 0.99)}
\usage{
label_pvalue(
  accuracy = 0.001,
  decimal.mark = NULL,
  prefix = NULL,
  add_p = FALSE
)
}
\arguments{
\item{accuracy}{A number to round to. Use (e.g.) \code{0.01} to show 2 decimal
places of precision. If \code{NULL}, the default, uses a heuristic that should
ensure breaks have the minimum number of digits needed to show the
difference between adjacent values.

Applied to rescaled data.}

\item{decimal.mark}{The character to be used to indicate the numeric
decimal point.  The default (\code{NULL}) retrieves the setting from the
\link[=number_options]{number options}.}

\item{prefix}{A character vector of length 3 giving the prefixes to
put in front of numbers. The default values are \code{c("p<", "p=", "p>")}
if \code{add_p} is \code{TRUE} and \code{c("<", "", ">")} if \code{FALSE}.}

\item{add_p}{Add "p=" before the value?}
}
\value{
All \code{label_()} functions return a "labelling" function, i.e. a function that
takes a vector \code{x} and returns a character vector of \code{length(x)} giving a
label for each input value.

Labelling functions are designed to be used with the \code{labels} argument of
ggplot2 scales. The examples demonstrate their use with x scales, but
they work similarly for all scales, including those that generate legends
rather than axes.
}
\description{
Formatter for p-values, using "<" and ">" for p-values close to 0 and 1.
}
\examples{
\dontshow{if (requireNamespace("ggplot2", quietly = TRUE)) withAutoprint(\{ # examplesIf}
library(ggplot2)
p <- sapply(mtcars[-1], function(x) cor.test(mtcars$mpg, x)$p.value)
cors <- data.frame(variable = names(p), p = p)
ggplot(cors, aes(x = p, y = variable)) +
  geom_col() +
  scale_x_continuous(labels = label_pvalue())
\dontshow{\}) # examplesIf}
demo_continuous(c(0, 1))
demo_continuous(c(0, 1), labels = label_pvalue())
demo_continuous(c(0, 1), labels = label_pvalue(accuracy = 0.1))
demo_continuous(c(0, 1), labels = label_pvalue(add_p = TRUE))

# Or provide your own prefixes
prefix <- c("p < ", "p = ", "p > ")
demo_continuous(c(0, 1), labels = label_pvalue(prefix = prefix))
}
\seealso{
Other labels for continuous scales:
\code{\link[=label_bytes]{label_bytes()}},
\code{\link[=label_currency]{label_currency()}},
\code{\link[=label_glue]{label_glue()}},
\code{\link[=label_number_auto]{label_number_auto()}},
\code{\link[=label_number_si]{label_number_si()}},
\code{\link[=label_ordinal]{label_ordinal()}},
\code{\link[=label_parse]{label_parse()}},
\code{\link[=label_percent]{label_percent()}},
\code{\link[=label_scientific]{label_scientific()}}
}
\concept{labels for continuous scales}
