% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-scientific.R
\name{label_scientific}
\alias{label_scientific}
\alias{scientific_format}
\alias{scientific}
\title{Label numbers with scientific notation}
\usage{
label_scientific(digits = 3, scale = 1, prefix = "", suffix = "",
  decimal.mark = ".", trim = TRUE, ...)

scientific_format(digits = 3, scale = 1, prefix = "", suffix = "",
  decimal.mark = ".", trim = TRUE, ...)

scientific(x, digits = 3, scale = 1, prefix = "", suffix = "",
  decimal.mark = ".", trim = TRUE, ...)
}
\arguments{
\item{digits}{Number of digits to show before exponent.}

\item{scale}{A scaling factor: \code{x} will be multiply by \code{scale} before
formating (useful if the underlying data is on another scale,
e.g. for computing percentages or thousands).}

\item{prefix, suffix}{Symbols to display before and after value.}

\item{decimal.mark}{The character to be used to indicate the numeric
decimal point.}

\item{trim}{Logical, if \code{FALSE}, values are right-justified to a common
width (see \code{\link[base:format]{base::format()}}).}

\item{...}{Other arguments passed on to \code{\link[base:format]{base::format()}}.}

\item{x}{A numeric vector to format.}
}
\value{
\code{*_format()} returns a function with single parameter
\code{x}, a numeric vector, that returns a character vector.
}
\description{
Label numbers with scientific notation
}
\section{Old interface}{

\code{scientific_format()} and \code{scientific()} are retired; please use
\code{label_scientific()}.
}

\examples{
demo_continuous(c(1, 10))
demo_continuous(c(1, 10), labels = label_scientific())
demo_continuous(c(1, 10), labels = label_scientific(digits = 3))
}
