% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-wrap.R
\name{label_wrap}
\alias{label_wrap}
\alias{wrap_format}
\title{Wrap long strings across multiple lines}
\usage{
label_wrap(width)

wrap_format(width)
}
\arguments{
\item{width}{Number of characters per line.}
}
\value{
\code{*_format()} returns a function with single parameter
\code{x}, a numeric vector, that returns a character vector.
}
\description{
Uses \code{\link[=strwrap]{strwrap()}} to split long labels across multiple lines.
}
\section{Old interface}{

\code{wrap_format()} is retired; please use \code{label_format()} instead.
}

\examples{
x <- c(
  "this is a long label",
  "this is another long label",
  "this a label this is even longer"
)
demo_discrete(x)
demo_discrete(x, labels = label_wrap(10))
demo_discrete(x, labels = label_wrap(20))
}
