% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trans-numeric.r
\name{log1p_trans}
\alias{log1p_trans}
\title{Log plus one transformation.}
\usage{
log1p_trans()
}
\description{
Log plus one transformation.
}
\examples{
trans_range(log_trans(), 1:10)
trans_range(log1p_trans(), 0:9)
}

