% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.r
\name{log_breaks}
\alias{log_breaks}
\title{Log breaks (integer breaks on log-transformed scales).}
\usage{
log_breaks(n = 5, base = 10)
}
\arguments{
\item{n}{desired number of breaks}

\item{base}{base of logarithm to use}
}
\description{
Log breaks (integer breaks on log-transformed scales).
}
\examples{
log_breaks()(c(1, 1e6))
log_breaks()(c(1, 1e5))
}

