% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trans-numeric.r
\name{log_trans}
\alias{log10_trans}
\alias{log2_trans}
\alias{log_trans}
\title{Log transformation.}
\usage{
log_trans(base = exp(1))
}
\arguments{
\item{base}{base of logarithm}
}
\description{
Log transformation.
}

