% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minor_breaks.R
\name{minor_breaks_width}
\alias{minor_breaks_width}
\alias{minor_breaks_n}
\title{Minor breaks}
\usage{
minor_breaks_width(width, offset)

minor_breaks_n(n)
}
\arguments{
\item{width}{Distance between each break. Either a number, or for
date/times, a single string of the form "{n} {unit}", e.g. "1 month",
"5 days". Unit can be of one "sec", "min", "hour", "day", "week",
"month", "year".}

\item{offset}{Use if you don't want breaks to start at zero}

\item{n}{number of breaks}
}
\description{
Generate minor breaks between major breaks either spaced with a fixed width,
or having a fixed number.
}
\examples{
demo_log10(c(1, 1e6))
if (FALSE) {
  # Requires https://github.com/tidyverse/ggplot2/pull/3591
  demo_log10(c(1, 1e6), minor_breaks = minor_breaks_n(10))
}
}
