\name{new_trans}
\alias{new_trans}
\title{Create a new transformation function...}
\usage{new_trans(name, transform, inverse, labels=inverse)}

\description{
  Create a new transformation function
}
\arguments{
  \item{name}{transformation name}
  \item{transform}{function, or name of function, that performs the}
  \item{inverse}{function, or name of function, that performs the}
}
