% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-auto.R
\name{number_auto}
\alias{number_auto}
\title{Automatic number formatter}
\usage{
number_auto(x)
}
\arguments{
\item{x}{A numeric vector to format.}
}
\description{
Switches between \code{\link[=comma]{comma()}} and \code{\link[=scientific]{scientific()}} format based on a set of
heuristics designed to automatically generate useful labels across a wide
range of inputs
}
\examples{
b <- function(a, b) extended_breaks(5)(c(a, b))

number_auto(b(0, 1e-6))
number_auto(b(0, 1e-3))
number_auto(b(0, 1))
number_auto(b(0, 1e3))
number_auto(b(0, 1e6))
number_auto(b(0, 1e9))
}
