% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels-retired.R
\name{number_bytes_format}
\alias{number_bytes_format}
\alias{number_bytes}
\title{Older interface to \code{label_bytes()}}
\usage{
number_bytes_format(symbol = "auto", units = "binary", ...)

number_bytes(x, symbol = "auto", units = c("binary", "si"), accuracy = 1, ...)
}
\arguments{
\item{symbol}{byte symbol to use. If "auto" the symbol used will be
determined separately for each value of \code{x}. Valid symbols are "B", "kB",
"MB", "GB", "TB", "PB", "EB", "ZB", and "YB" for SI units, and the "iB"
variants for binary units.}

\item{units}{which unit base to use, "binary" (1024 base) or "si" (1000 base)}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("retired")}
These functions are kept for backward compatibility, but you should switch
to \code{\link[=label_bytes]{label_bytes()}} for new code.
}
\keyword{internal}
