% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-ordinal.R
\name{ordinal_format}
\alias{ordinal_format}
\alias{ordinal}
\title{Superseded interface to \code{label_ordinal()}}
\usage{
ordinal_format(
  prefix = "",
  suffix = "",
  big.mark = " ",
  rules = ordinal_english(),
  ...
)

ordinal(
  x,
  prefix = "",
  suffix = "",
  big.mark = " ",
  rules = ordinal_english(),
  ...
)
}
\arguments{
\item{prefix, suffix}{Symbols to display before and after value.}

\item{big.mark}{Character used between every 3 digits to separate thousands.}

\item{rules}{Named list of regular expressions, matched in order.
Name gives suffix, and value specifies which numbers to match.}

\item{...}{Other arguments passed on to \code{\link[base:format]{base::format()}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

These functions are kept for backward compatibility; you should switch
to \code{\link[=label_ordinal]{label_ordinal()}} for new code.
}
\keyword{internal}
