% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formatter.r
\name{ordinal_format}
\alias{ordinal}
\alias{ordinal_format}
\title{Ordinal formatter: add ordinal suffixes (-st, -nd, -rd, -th) to numbers.}
\usage{
ordinal_format(x)

ordinal(x)
}
\arguments{
\item{x}{a numeric vector to format}
}
\value{
a function with single paramater x, a numeric vector, that
  returns a character vector
}
\description{
Ordinal formatter: add ordinal suffixes (-st, -nd, -rd, -th) to numbers.
}
\examples{
ordinal_format()(1:10)
ordinal(1:10)
}

