% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-expression.R
\name{parse_format}
\alias{parse_format}
\alias{math_format}
\title{Label with mathematical expressions}
\usage{
parse_format()

math_format(expr = 10^.x, format = force)
}
\arguments{
\item{expr}{expression to use}

\item{format}{another format function to apply prior to mathematical
transformation - this makes it easier to use floating point numbers in
mathematical expressions.}
}
\value{
A function with a single parameter, \code{x}, that returns an
expression object.
}
\description{
\code{parse_format()} produces expression from strings by parsing them;
\code{math_format()} constructs expressions by replacing the pronoun \code{.x}
with each string.
}
\examples{
# Use parse_format() with discrete scales
greek <- c("alpha", "beta", "gamma")
demo_discrete(greek)
demo_discrete(greek, labels = parse_format())

# Use math_format() with continuous scales
demo_continuous(c(1, 5))
demo_continuous(c(1, 5), labels = math_format(alpha[.x]))
}
\seealso{
\link{plotmath} for the details of mathematical formatting in R.
}
