\name{parse_format}
\alias{parse_format}
\title{Parse a text label to produce expressions for plotmath.}

\description{
  Parse a text label to produce expressions for plotmath.
}
\seealso{\code{\link{plotmath}}}
\value{a function with single paramater x, a character vector, that
returns a list of expressions}
\arguments{
  \item{x}{character vector to format}
}
