% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-expression.R
\name{parse_format}
\alias{parse_format}
\alias{math_format}
\title{Superseded interface to \code{label_parse()}/\code{label_math()}}
\usage{
parse_format()

math_format(expr = 10^.x, format = force)
}
\arguments{
\item{expr}{expression to use}

\item{format}{another format function to apply prior to mathematical
transformation - this makes it easier to use floating point numbers in
mathematical expressions.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

These functions are kept for backward compatibility; you should switch
to \code{\link[=label_parse]{label_parse()}}/\code{\link[=label_math]{label_math()}} for new code.
}
\keyword{internal}
