% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formatter.r
\name{parse_format}
\alias{parse_format}
\title{Parse a text label to produce expressions for plotmath.}
\usage{
parse_format()
}
\value{
a function with single parameter x, a character vector, that
   returns a list of expressions
}
\description{
Parse a text label to produce expressions for plotmath.
}
\examples{
parse_format()(c("alpha", "beta", "gamma"))
}
\seealso{
\code{\link{plotmath}}
}

