\name{percent_format}
\alias{percent_format}
\title{Percent formatter: multiply by one hundred and display percent sign.}

\description{
  Percent formatter: multiply by one hundred and display
  percent sign.
}
\value{a function with single paramater x, a numeric vector, that
returns a character vector}
