\name{pretty_breaks}
\alias{pretty_breaks}
\title{Pretty breaks.}
\usage{pretty_breaks(n=5, ...)}

\description{
  Pretty breaks. Uses default R break algorithm as
  implemented in \code{\link{pretty}}
}
\arguments{
  \item{n}{desired number of breaks}
  \item{...}{other arguments passed on to \code{\link{pretty}}}
}
