% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{pretty_breaks}
\alias{pretty_breaks}
\title{Superseded interface to \code{breaks_pretty()}}
\usage{
pretty_breaks(n = 5, ...)
}
\arguments{
\item{n}{Desired number of breaks. You may get slightly more or fewer
breaks that requested.}

\item{...}{other arguments passed on to \code{\link[=pretty]{pretty()}}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

These functions are kept for backward compatibility; you should switch
to \code{\link[=breaks_pretty]{breaks_pretty()}} for new code.
}
\keyword{internal}
