% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-pvalue.R
\name{pvalue_format}
\alias{pvalue_format}
\alias{pvalue}
\title{Superseded interface to \code{label_pvalue()}}
\usage{
pvalue_format(
  accuracy = 0.001,
  decimal.mark = ".",
  prefix = NULL,
  add_p = FALSE
)

pvalue(x, accuracy = 0.001, decimal.mark = ".", prefix = NULL, add_p = FALSE)
}
\arguments{
\item{accuracy}{A number to round to. Use (e.g.) \code{0.01} to show 2 decimal
places of precision. If \code{NULL}, the default, uses a heuristic that should
ensure breaks have the minimum number of digits needed to show the
difference between adjacent values.

Applied to rescaled data.}

\item{decimal.mark}{The character to be used to indicate the numeric
decimal point.}

\item{prefix}{A character vector of length 3 giving the prefixes to
put in front of numbers. The default values are \code{c("<", "", ">")}
if \code{add_p} is \code{TRUE} and \code{c("p<", "p=", "p>")} if \code{FALSE}.}

\item{add_p}{Add "p=" before the value?}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

These functions are kept for backward compatibility; you should switch
to \code{\link[=label_pvalue]{label_pvalue()}} for new code.
}
\keyword{internal}
