% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.r
\docType{class}
\name{Range-class}
\alias{ContinuousRange}
\alias{DiscreteRange}
\alias{Range}
\alias{Range-class}
\title{Mutable ranges.}
\description{
Mutable ranges have a two methods (\code{train} and \code{reset}), and
make it possible to build up complete ranges with multiple passes.
}

