% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.r
\docType{data}
\name{Range}
\alias{Range}
\alias{DiscreteRange}
\alias{ContinuousRange}
\title{Mutable ranges}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Range

DiscreteRange

ContinuousRange
}
\description{
Mutable ranges have a two methods (\code{train} and \code{reset}), and
make it possible to build up complete ranges with multiple passes.
}
\keyword{datasets}
