% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.r
\name{rescale_none}
\alias{rescale_none}
\title{Don't peform rescaling}
\usage{
rescale_none(x, ...)
}
\arguments{
\item{x}{numeric vector of values to manipulate.}

\item{...}{all other arguments ignored}
}
\description{
Don't peform rescaling
}
\examples{
rescale_none(1:100)
}
