% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-rescale.r
\name{rescale_pal}
\alias{rescale_pal}
\title{Rescale palette (continuous).}
\usage{
rescale_pal(range = c(0.1, 1))
}
\arguments{
\item{range}{Numeric vector of length two, giving range of possible
values.  Should be between 0 and 1.}
}
\description{
Just rescales the input to the specific output range.  Useful for
alpha, size, and continuous position.
}

