% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales-package.R
\docType{package}
\name{scales-package}
\alias{scales}
\alias{scales-package}
\title{scales: Scale Functions for Visualization}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Graphical scales map data to aesthetics, and provide methods for automatically determining breaks and labels for axes and legends.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://scales.r-lib.org}
  \item \url{https://github.com/r-lib/scales}
  \item Report bugs at \url{https://github.com/r-lib/scales/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomas.pedersen@posit.co} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})

Authors:
\itemize{
  \item Thomas Lin Pedersen \email{thomas.pedersen@posit.co} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})
  \item Hadley Wickham \email{hadley@posit.co}
  \item Dana Seidel
}

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
