% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales-package.R
\docType{package}
\name{scales-package}
\alias{scales}
\alias{scales-package}
\title{scales: Scale Functions for Visualization}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Graphical scales map data to aesthetics, and provide methods for automatically determining breaks and labels for axes and legends.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://scales.r-lib.org}
  \item \url{https://github.com/r-lib/scales}
  \item Report bugs at \url{https://github.com/r-lib/scales/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Authors:
\itemize{
  \item Dana Seidel
}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
