% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour-manip.R
\name{show_col}
\alias{show_col}
\title{Show colours}
\usage{
show_col(colours, labels = TRUE, borders = NULL, cex_label = 1, ncol = NULL)
}
\arguments{
\item{colours}{A character vector of colours}

\item{labels}{Label each colour with its hex name?}

\item{borders}{Border colour for each tile. Default uses \code{par("fg")}.
Use \code{border = NA} to omit borders.}

\item{cex_label}{Size of printed labels, as multiplier of default size.}

\item{ncol}{Number of columns. If not supplied, tries to be as square as
possible.}
}
\description{
A quick and dirty way to show colours in a plot.
}
\examples{
show_col(pal_hue()(9))
show_col(pal_hue()(9), borders = NA)

show_col(pal_viridis()(16))
show_col(pal_viridis()(16), labels = FALSE)
}
\keyword{internal}
