% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour-manip.r
\name{show_col}
\alias{show_col}
\title{Show colours}
\usage{
show_col(colours, labels = TRUE, borders = NULL, cex_label = 1)
}
\arguments{
\item{colours}{a character vector of colours}

\item{labels}{boolean, whether to show the hexadecimal representation of the colours in each tile}

\item{borders}{colour of the borders of the tiles; matches the \code{border} argument of \code{\link[graphics:rect]{graphics::rect()}}. The default means \code{par("fg")}. Use \code{border = NA} to omit borders.}

\item{cex_label}{size of printed labels, works the same as \code{cex} parameter of \code{plot()}}
}
\description{
A quick and dirty way to show colours in a plot.
}
\keyword{internal}
