% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-continuous.r
\name{train_continuous}
\alias{train_continuous}
\title{Train (update) a continuous scale}
\usage{
train_continuous(new, existing = NULL)
}
\arguments{
\item{new}{New data to add to scale}

\item{existing}{Optional existing scale to update}
}
\description{
Train (update) a continuous scale
}
