% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scale-discrete.r
\name{train_discrete}
\alias{train_discrete}
\title{Train (update) a discrete scale}
\usage{
train_discrete(new, existing = NULL, drop = FALSE)
}
\arguments{
\item{new}{New data to add to scale}

\item{existing}{Optional existing scale to update}

\item{drop}{\code{TRUE}, will drop factor levels not associated with data}
}
\description{
Train (update) a discrete scale
}

