% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels-retired.R
\name{trans_format}
\alias{trans_format}
\title{Format labels after transformation}
\usage{
trans_format(trans, format = scientific_format())
}
\arguments{
\item{trans}{transformation to apply}

\item{format}{additional formatter to apply after transformation}
}
\value{
a function with single parameter x, a numeric vector, that
returns a character vector of list of expressions
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
}
\examples{
tf <- trans_format("log10", scientific_format())
tf(10^1:6)
}
\keyword{internal}
