% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.r
\name{trans_range}
\alias{trans_range}
\title{Compute range of transformed values.}
\usage{
trans_range(trans, x)
}
\arguments{
\item{trans}{a transformation object, or the name of a transformation object
given as a string.}

\item{x}{a numeric vector to compute the rande of}
}
\description{
Silently drops any ranges outside of the domain of \code{trans}.
}
