% Generated by roxygen2 (4.0.1): do not edit by hand
\name{unit_format}
\alias{unit_format}
\title{Add units to the labels}
\usage{
unit_format(unit = "m", scale = 1, sep = " ", ...)
}
\arguments{
\item{unit}{The units to append}

\item{scale}{A scaling factor. Useful if the underlying data is on another scale}

\item{sep}{The separator between the number and the label}

\item{...}{Arguments passed to \code{\link{comma}}}
}
\description{
Add units to the labels
}
\examples{
# labels in kilometer when the raw data are in meter
  unit_format(unit = "km", scale = 1e-3)
  # labels in hectares, raw data in square meters
  unit_format(unit = "ha", scale = 1e-4)
}
\seealso{
\code{\link{comma}}
}

