% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-wrap.R
\name{wrap_format}
\alias{wrap_format}
\title{Superseded interface to \code{label_wrap()}}
\usage{
wrap_format(width)
}
\arguments{
\item{width}{Number of characters per line.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

These functions are kept for backward compatibility; you should switch
to \code{\link[=label_wrap]{label_wrap()}} for new code.
}
\keyword{internal}
