% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatter.r
\name{wrap_format}
\alias{wrap_format}
\title{Wrap text to a specified width, adding newlines for spaces if text exceeds
the width}
\usage{
wrap_format(width)
}
\arguments{
\item{width}{value above which to wrap}
}
\value{
Function with single parameter x, a character vector, that
   returns a wrapped character vector
}
\description{
Wrap text to a specified width, adding newlines for spaces if text exceeds
the width
}
\examples{
wrap_10 <- wrap_format(10)
wrap_10('A long line that needs to be wrapped')
}

